v2.1.0 (#5)
* Make workflow separate `develop` from `release` tags for artifact builds

* Add comments and replace `console` log

* Add contributing guidelines for the repository (#2)

* Make Release Drafter follow `develop` branch

* Add CONTRIBUTING.md and update README.md

* Results pagination overhaul (#3)

* Replace `offset` parameter with `page` for pagination of results

Instead of taking the offset directly from the user query, we can
compute it by multiplying the page number with the `limit` parameter
set in the query or by default.

Not having a `page` parameter in the query is equivalent to `page=1`.

* Fix `limit <= 0` and add missing tests

* Add `parsePage` tests

* Add metadata information to API responses

The metadata contains information for the pagination of results.
The following fields are available:
- `next_page`, capped to `total_pages` on reaching the last page
- `previous_page`, set to 1 on the first page
- `total_pages`, computed from the `limit` parameter
- `total_results` returned from `rows_before_limit_at_least` of
   ClickHouse's response
   (https://clickhouse.com/docs/en/interfaces/formats#json)

* Make `limit` parameter default to `config.maxLimit`

Resolves #1

* Edit automation and documentation to drop `develop` (#4)

Given the relatively small size of the repo, it will be easier to use
to make PRs directly to the `main` branch, keeping the current tagging
schema intact.

PRs would still be required before merging to `main` following the
conventions described in `CONTRIBUTING.md`.