Fix pagination pages being parsed as strings
The `limit` and `page` parameters were transformed as `ZodDefault` after
the first query. This means the routine for identifying the underlying
type would fail as it didn't fall in the `ZodOptional` or `ZodUnion` cases.

Hence, those parameters would be interpreted as strings and the `+` operator
for the calculation of pages would be interpreted as concatenation rather
than integer addition.